feature(metrics): optimizing metrics to relieve gc presure

+++ pkg/typedef/typedef.go
@@ -72,7 +72,7 @@ func (q SimpleQuery) ToCql() (stmt string, names []string) {
 type (
 	PartitionKeys struct {
 		Values Values
-		Token  uint64
+		Token  uint32
 	}
 
 	Stmt struct {
